separate light protocol handler from provider
